consolidate files into btgui/OpenGLWindow (from FontFiles/OpenGLTrueTypeFont) consolidate stringify_linux/_osx into stringify.sh
