+++ Demos3/CpuDemos/main_opengl3core.cpp
@@ -22,8 +22,8 @@
 
 #include <assert.h>
 #include <string.h>
-#include "OpenGLTrueTypeFont/fontstash.h"
-#include "OpenGLTrueTypeFont/opengl_fontstashcallbacks.h"
+#include "OpenGLWindow/fontstash.h"
+#include "OpenGLWindow/opengl_fontstashcallbacks.h"
 #include "gwenUserInterface.h"
 
 #include "../btgui/Timing/b3Quickprof.h"
@@ -459,7 +459,7 @@ void myprintf(const char* msg)
 
 
 //#define STB_IMAGE_WRITE_IMPLEMENTATION
-#include "OpenGLTrueTypeFont/stb_image_write.h"
+#include "OpenGLWindow/stb_image_write.h"
 void writeTextureToPng(int textureWidth, int textureHeight, const char* fileName)
 {
 	int numComponents = 4;
