+++ Demos3/GpuDemos/main_opengl3core.cpp
@@ -24,8 +24,8 @@
 //#include "b3GpuDynamicsWorld.h"
 #include <assert.h>
 #include <string.h>
-#include "OpenGLTrueTypeFont/fontstash.h"
-#include "OpenGLTrueTypeFont/opengl_fontstashcallbacks.h"
+#include "OpenGLWindow/fontstash.h"
+#include "OpenGLWindow/opengl_fontstashcallbacks.h"
 #include "gwenUserInterface.h"
 #include "ParticleDemo.h"
 #include "broadphase/PairBench.h"
@@ -543,7 +543,7 @@ void myprintf(const char* msg)
 
 
 //#define STB_IMAGE_WRITE_IMPLEMENTATION
-#include "OpenGLTrueTypeFont/stb_image_write.h"
+#include "OpenGLWindow/stb_image_write.h"
 void writeTextureToPng(int textureWidth, int textureHeight, const char* fileName)
 {
 	int numComponents = 4;
