+++ Demos3/GpuDemos/premake4.lua
@@ -52,11 +52,11 @@ function createProject(vendor)
 			"../../btgui/OpenGLWindow/TwFonts.h",
 			"../../btgui/OpenGLWindow/GLRenderToTexture.cpp",
 			"../../btgui/OpenGLWindow/GLRenderToTexture.h",
-			"../../btgui/OpenGLTrueTypeFont/fontstash.cpp",
-			"../../btgui/OpenGLTrueTypeFont/fontstash.h",
-			"../../btgui/OpenGLTrueTypeFont/opengl_fontstashcallbacks.cpp",
-			"../../btgui/OpenGLTrueTypeFont/opengl_fontstashcallbacks.h",
-			"../../btgui/FontFiles/OpenSans.cpp",
+			"../../btgui/OpenGLWindow/fontstash.cpp",
+			"../../btgui/OpenGLWindow/fontstash.h",
+			"../../btgui/OpenGLWindow/opengl_fontstashcallbacks.cpp",
+			"../../btgui/OpenGLWindow/opengl_fontstashcallbacks.h",
+			"../../btgui/OpenGLWindow/OpenSans.cpp",
 			"../../btgui/stb_image/stb_image.cpp",
 			"../../btgui/stb_image/stb_image.h",
 			"../../btgui/Timing/b3Quickprof.cpp",
