+++ Demos3/GpuGuiInitialize/main.cpp
@@ -14,8 +14,8 @@
 //#include "b3GpuDynamicsWorld.h"
 #include <assert.h>
 #include <string.h>
-#include "OpenGLTrueTypeFont/fontstash.h"
-#include "OpenGLTrueTypeFont/opengl_fontstashcallbacks.h"
+#include "OpenGLWindow/fontstash.h"
+#include "OpenGLWindow/opengl_fontstashcallbacks.h"
 
 #include "OpenGLWindow/GwenOpenGL3CoreRenderer.h"
 #include "../btgui/Timing/b3Quickprof.h"
