+++ Demos3/GpuGuiInitialize/premake4.lua
@@ -42,11 +42,11 @@ function createProject(vendor)
 			"../../btgui/OpenGLWindow/TwFonts.cpp",
 			"../../btgui/OpenGLWindow/TwFonts.h",
       "../../btgui/OpenGLWindow/GwenOpenGL3CoreRenderer.h",
-      "../../btgui/FontFiles/OpenSans.cpp",
-			"../../btgui/OpenGLTrueTypeFont/fontstash.cpp",
-			"../../btgui/OpenGLTrueTypeFont/fontstash.h",
-			"../../btgui/OpenGLTrueTypeFont/opengl_fontstashcallbacks.cpp",
- 			"../../btgui/OpenGLTrueTypeFont/opengl_fontstashcallbacks.h",
+      "../../btgui/OpenGLWindow/OpenSans.cpp",
+			"../../btgui/OpenGLWindow/fontstash.cpp",
+			"../../btgui/OpenGLWindow/fontstash.h",
+			"../../btgui/OpenGLWindow/opengl_fontstashcallbacks.cpp",
+ 			"../../btgui/OpenGLWindow/opengl_fontstashcallbacks.h",
 			"../../src/Bullet3Geometry/b3ConvexHullComputer.cpp",
 			"../../src/Bullet3Geometry/b3ConvexHullComputer.h",
 			"../../src/Bullet3Common/b3AlignedAllocator.cpp",
