+++ build3/bin2cpp.bat
@@ -2,6 +2,6 @@
 rem @echo off
 
 
-premake4 --file=bin2cpp.lua --binfile="../btgui/FontFiles/OpenSans.ttf" --cppfile="../btgui/FontFiles/OpenSans.cpp" --stringname="OpenSansData" bin2cpp
+premake4 --file=bin2cpp.lua --binfile="../btgui/FontFiles/OpenSans.ttf" --cppfile="../btgui/OpenGLWindow/OpenSans.cpp" --stringname="OpenSansData" bin2cpp
 
 pause
