+++ build3/stringify.sh
